[Fix] export war build.gradle, CommandRunner.java 메서드를 AnitimeApplication.java에서 분리(외장톸캣issue)
